move remaining Fast*-router classes to locationchoice contrib
the locationchoice is now the only place where some customized fast-routing algorithms are used.
