+++ contribs/locationchoice/src/main/java/org/matsim/core/router/MultiNodeDijkstra.java
@@ -43,66 +43,66 @@ import org.matsim.vehicles.Vehicle;
 /**
  * <p>An extended implementation of the Dijkstra algorithm that supports multiple
  * start- and/or end nodes.</p>
- * 
+ *
  * <p>To do so, ImaginaryNodes and IntialNodes are introduced which is required for
  * backwards compatibility with the default Dijkstra implementation which expects
  * only single nodes as arguments.</p>
- * 
+ *
  * <p>Therefore, ImaginaryNodes are introduced. They contain a collection of
  * InitialNodes that represents the start or end nodes to be used by the algorithm.
  * Each InitialNode represents a node from the network. In addition, initial time and
  * costs can be defined, i.e. the time respectively costs to reach that node.<p>
- * 
+ *
  * <p>By default, only the cheapest path between is calculated, i.e. the routing algorithm
  * will terminate before all end nodes have been reached. This behaviour can be changed
  * by setting the searchAllEndNodes parameter to true.</p>
- * 
+ *
  * @see org.matsim.core.router.Dijkstra
  * @see org.matsim.core.router.InitialNode
- * @see org.matsim.core.router.ImaginaryNode
- * 
+ * @see ImaginaryNode
+ *
  * @author cdobler
  */
 public class MultiNodeDijkstra extends Dijkstra implements MultiNodePathCalculator {
-	
+
 	private final static Logger log = LogManager.getLogger(MultiNodeDijkstra.class);
-	
+
 	/*
 	 * If this value is true, the algorithm tries to find routes to all end nodes.
 	 * Otherwise only the one(s) with the lowest costs are found.
-	 * 
+	 *
 	 * When enabling this option, select end nodes with care! If only one of them
 	 * is in a not reachable part of the network, the algorithm will process the
 	 * entire reachable part of the network before terminating.
 	 */
 	private boolean searchAllEndNodes;
-	
+
 	MultiNodeDijkstra(Network network, TravelDisutility costFunction, TravelTime timeFunction, boolean searchAllEndNodes) {
 		super(network, costFunction, timeFunction);
 		this.searchAllEndNodes = searchAllEndNodes;
 	}
-	
+
 	MultiNodeDijkstra(final Network network, final TravelDisutility costFunction, final TravelTime timeFunction,
 			final PreProcessDijkstra preProcessData, boolean searchAllEndNodes) {
 		super(network, costFunction, timeFunction, preProcessData);
 		this.searchAllEndNodes = searchAllEndNodes;
 	}
-	
+
 	public static ImaginaryNode createImaginaryNode(Collection<? extends InitialNode> nodes) {
 		return new ImaginaryNode(nodes);
 	}
-	
+
 	public static ImaginaryNode createImaginaryNode(Collection<? extends InitialNode> nodes, Coord coord) {
 		return new ImaginaryNode(nodes, coord);
 	}
-	
+
 	/*
 	 * We have to extend this method from the original Dijkstra. The given input nodes might be
 	 * ImaginaryNodes which contain multiple start / end nodes for the routing process. Those
 	 * nodes should be part of the routing network, therefore we perform the check for each of them.
-	 * 
+	 *
 	 * Regular nodes are directly passed to the super class.
-	 * 
+	 *
 	 * cdobler, jun'14
 	 */
 	/*package*/ @Override
@@ -112,29 +112,29 @@ public class MultiNodeDijkstra extends Dijkstra implements MultiNodePathCalculat
 			for (InitialNode initialNode : imaginaryNode.initialNodes) super.checkNodeBelongToNetwork(initialNode.node);
 		} else super.checkNodeBelongToNetwork(node);
 	}
-	
+
 	@Override
 	/*package*/ Node searchLogic( final Node fromNode, final Node toNode, final RouterPriorityQueue<Node> pendingNodes, Person person,
 				      Vehicle vehicle ) {
-		
+
 		// If it is an imaginary node...
 		if (toNode instanceof ImaginaryNode) {
-			
+
 			Map<Id<Node>, InitialNode> endNodes = new HashMap<>();
-			
+
 			Collection<? extends InitialNode> initialNodes = ((ImaginaryNode) toNode).initialNodes;
 			for (InitialNode initialNode : initialNodes) endNodes.put(initialNode.node.getId(), initialNode);
 
 			// find out which one is the cheapest end node
 			double minCost = Double.POSITIVE_INFINITY;
 			Node minCostNode = null;
-			
+
 			// continue searching as long as unvisited end nodes are available
 			boolean stillSearching = endNodes.size() > 0;
-			
+
 			while (stillSearching) {
 				Node outNode = pendingNodes.poll();
-				
+
 				if (outNode == null) {
 					/*
 					 * This is not necessarily a problem. Some of the out nodes might be reachable only
@@ -154,24 +154,24 @@ public class MultiNodeDijkstra extends Dijkstra implements MultiNodePathCalculat
 							log.trace(sb.toString());
 						}
 					}
-					
+
 					if (searchAllEndNodes && endNodes.size() > 0) {
 						for (InitialNode endNode : endNodes.values()) {
 							log.trace("No route was found from node " + fromNode.getId() + " to destination node " + endNode.node.getId() + ".");
 							log.trace( Dijkstra.createInfoMessage( person, vehicle ) );
 						}
 					}
-					
+
 					endNodes.clear();
 					stillSearching = false;
 				} else {
 					DijkstraNodeData data = getData(outNode);
 					InitialNode initData = endNodes.remove(outNode.getId());
-										
+
 					/*
 					 * If the node is an end node.
-					 * Note: 
-					 * The node was head of the priority queue, i.e. the shortest path to the 
+					 * Note:
+					 * The node was head of the priority queue, i.e. the shortest path to the
 					 * node has been found. The algorithm will not re-visit the node on another
 					 * route!
 					 */
@@ -182,7 +182,7 @@ public class MultiNodeDijkstra extends Dijkstra implements MultiNodePathCalculat
 							minCostNode = outNode;
 						}
 					}
-					
+
 					if (searchAllEndNodes) {
 						relaxNode(outNode, null, pendingNodes);
 						stillSearching = endNodes.size() > 0;
@@ -196,13 +196,13 @@ public class MultiNodeDijkstra extends Dijkstra implements MultiNodePathCalculat
 					}
 				}
 			}
-			
+
 			return minCostNode;
-		} 
+		}
 		// ... otherwise: default behaviour.
 		else return super.searchLogic(fromNode, toNode, pendingNodes, person, vehicle );
 	}
-	
+
 	/*
 	 * initFromNode -> first Node in pendingNodes -> calcLeastCostPath will relax that node
 	 * -> we relax that dummy node here
@@ -210,18 +210,18 @@ public class MultiNodeDijkstra extends Dijkstra implements MultiNodePathCalculat
 	@Override
 	/*package*/ void initFromNode(final Node fromNode, final Node toNode, final double startTime,
 			final RouterPriorityQueue<Node> pendingNodes) {
-		
+
 		// If it is an imaginary node, we relax it.
-		if (fromNode instanceof ImaginaryNode) {			
+		if (fromNode instanceof ImaginaryNode) {
 			relaxImaginaryNode((ImaginaryNode) fromNode, pendingNodes, startTime);
 		}
 		// ... otherwise: default behaviour.
 		else super.initFromNode(fromNode, toNode, startTime, pendingNodes);
 	}
-	
+
 	protected void relaxImaginaryNode(final ImaginaryNode outNode, final RouterPriorityQueue<Node> pendingNodes,
 			final double currTime) {
-		
+
 		double currCost = 0.0;	// should be 0
 
 		for (InitialNode initialNode : outNode.initialNodes) {
@@ -229,10 +229,10 @@ public class MultiNodeDijkstra extends Dijkstra implements MultiNodePathCalculat
 			double travelCost = initialNode.initialCost;
 			DijkstraNodeData data = getData(initialNode.node);
 			Link l = null;	// fromLink - use a dummy link here??
-			visitNode(initialNode.node, data, pendingNodes, currTime + travelTime, currCost + travelCost, l);			
+			visitNode(initialNode.node, data, pendingNodes, currTime + travelTime, currCost + travelCost, l);
 		}
-	}	
-	
+	}
+
 	@Override
 	protected Path constructPath(Node fromNode, Node toNode, double startTime, double arrivalTime) {
 		ArrayList<Node> nodes = new ArrayList<Node>();
@@ -240,7 +240,7 @@ public class MultiNodeDijkstra extends Dijkstra implements MultiNodePathCalculat
 
 		nodes.add(0, toNode);
 		Link tmpLink = getData(toNode).getPrevLink();
-		
+
 		// Only this part has been adapted. Could probably also be changed in the super class.
 		while (tmpLink != null) {
 			links.add(0, tmpLink);
@@ -256,12 +256,12 @@ public class MultiNodeDijkstra extends Dijkstra implements MultiNodePathCalculat
 
 		return path;
 	}
-	
+
 	/**
 	 * This method should only be called from outside after calcLeastCostPath(...) has
 	 * been executed. After that, the paths between the multiple start and/or end nodes
 	 * can be constructed using this method.
-	 * 
+	 *
 	 * Is there a way to check whether this method is called as intended??
 	 * cdobler, oct'13
 	 */
@@ -275,9 +275,9 @@ public class MultiNodeDijkstra extends Dijkstra implements MultiNodePathCalculat
 			if (!fromData.isVisited(this.getIterationId())) return null;
 
 			double arrivalTime = toData.getTime();
-			
+
 			// now construct and return the path
-			return constructPath(fromNode, toNode, startTime, arrivalTime);			
+			return constructPath(fromNode, toNode, startTime, arrivalTime);
 		}
 	}
