+++ contribs/locationchoice/src/main/java/org/matsim/core/router/MultiNodePathCalculator.java
@@ -21,11 +21,12 @@
 package org.matsim.core.router;
 
 import org.matsim.api.core.v01.network.Node;
+import org.matsim.core.router.ImaginaryNode;
 import org.matsim.core.router.util.LeastCostPathCalculator;
 
 /**
  * Marker interface so that one can program against an interface rather than against a very specific implementation.
- * 
+ *
  * @author nagel
  */
 public interface MultiNodePathCalculator extends LeastCostPathCalculator {
