+++ contribs/locationchoice/src/main/java/org/matsim/core/router/util/AbstractRoutingNetworkFactory.java
@@ -1,10 +1,10 @@
 /* *********************************************************************** *
  * project: org.matsim.*
- * RoutingNetworkLink.java
+ * AbstractRoutingNetworkFactory.java
  *                                                                         *
  * *********************************************************************** *
  *                                                                         *
- * copyright       : (C) 2011 by the members listed in the COPYING,        *
+ * copyright       : (C) 2012 by the members listed in the COPYING,        *
  *                   LICENSE and WARRANTY file.                            *
  * email           : info at matsim dot org                                *
  *                                                                         *
@@ -20,25 +20,20 @@
 
 package org.matsim.core.router.util;
 
+import org.matsim.api.core.v01.Coord;
 import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.network.Link;
+import org.matsim.api.core.v01.network.Node;
 
-public interface RoutingNetworkLink extends Link {
-	// yy In most of matsim we have the convention that the wrapped object is provided, but the methods are not delegated.  I.e.
-	//    routingNetworkLink.getLink().getFreeSpeed() ;
-	// but NOT
-	//    routingNetworkLink.getFreeSpeed() ;
-	// The current approach seems to be doing both.  Why did it add the second way of doing things?  
-	// (I can imagine that this came from retrofitting, but a design choice explanation would still be helpful.)  kai, may'17
+public abstract class AbstractRoutingNetworkFactory implements RoutingNetworkFactory {
 
-	public Link getLink();
-	
 	@Override
-	public Id<Link> getId();
-	
+	public Link createLink(final Id<Link> id, final Node fromNode, final Node toNode) {
+		throw new RuntimeException("Not supported operation!");
+	}
+
 	@Override
-	public RoutingNetworkNode getFromNode();
-	
-	@Override
-	public RoutingNetworkNode getToNode();
+	public RoutingNetworkNode createNode(final Id<Node> id, final Coord coord) {
+		throw new RuntimeException("Not supported operation!");
+	}
 }
