+++ contribs/locationchoice/src/main/java/org/matsim/core/router/util/AbstractRoutingNetworkLink.java
