+++ matsim/src/main/java/org/matsim/core/router/priorityqueue/HasIndex.java
@@ -20,21 +20,17 @@
 
 package org.matsim.core.router.priorityqueue;
 
-import org.matsim.core.router.util.ArrayRoutingNetwork;
-import org.matsim.core.router.util.ArrayRoutingNetworkNode;
-
 /**
  * An interface to mark classes that enumerate their objects. Each index
  * should be unique and can e.g. be used to lookup values in an array. This feature
  * is used in some classed due to performance reasons since a lookup in an array
  * is much faster than in a map.
- * 
- * @see ArrayRoutingNetwork
- * @see ArrayRoutingNetworkNode
+ *
  * @see BinaryMinHeap
- * 
+ *
  * @author cdobler
  */
+@Deprecated // Id.index() should be used instead nowadays.
 public interface HasIndex {
 
 	public int getArrayIndex();
